backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
pydantic>=2.4.2
uvicorn>=0.23.2
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.0

# Date/time handling
pytz>=2023.3

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Type checking
mypy>=1.6.1

# Clerk SDK
clerk-python

# Google Auth
google-auth-oauthlib
google-api-python-client